+++ public/admin/_app/immutable/nodes/24.BGPK5TIc.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/e82hilsr.js";export{e as component};